pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "weather"
version = "0.0.1"

authors = [
  { name="George Chaklyarov", email="chaklyarov@gmail.com" },
]
description = "A small example package"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "spark",
    "pandas"
]

[project.urls]
"Homepage" = "https://apmath.spbu.ru"
"Bug Tracker" = "https://github.com/4eckah78/typical_python_project"

[tool.isort]
profile = "black"
src_paths = ["projectname", "test"]

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = false
disallow_incomplete_defs = false
follow_imports = "normal"
ignore_missing_imports = true
pretty = true
show_column_numbers = true
show_error_codes = true
warn_no_return = false
warn_unused_ignores = true

[tool.black]
skip-string-normalization = true

[tool.ruff]
line-length = 88